+++ src/Netnr.ClassLibrary/Netnr.Fast/DownTo.cs
@@ -9,7 +9,7 @@ namespace Netnr.Fast
     /// </summary>
     public class DownTo
     {
-        private HttpResponse Response;
+        private readonly HttpResponse Response;
 
         /// <summary>
         /// 构造
@@ -30,12 +30,22 @@ namespace Netnr.Fast
             fileStream.Read(bytes, 0, bytes.Length);
             fileStream.Close();
 
+            Stream(bytes, fileName);
+        }
+
+        /// <summary>
+        /// 流的方式下载
+        /// </summary>
+        /// <param name="bytes"></param>
+        /// <param name="fileName">下载文件名</param>
+        public void Stream(byte[] bytes, string fileName)
+        {
             Response.ContentType = "application/octet-stream";
 
             // 通知浏览器下载而不是打开  
-                Response.Headers.Add("Content-Disposition", "attachment; filename=" + HttpUtility.UrlEncode(fileName, System.Text.Encoding.UTF8));
-            Response.Body.Write(bytes, 0, bytes.Length);
-            Response.Body.Flush();
+            Response.Headers.Add("Content-Disposition", "attachment; filename=" + HttpUtility.UrlEncode(fileName, System.Text.Encoding.UTF8));
+            Response.Body.WriteAsync(bytes, 0, bytes.Length).Wait();
+            Response.Body.FlushAsync().Wait();
         }
     }
 }